+++ Transportation/Entities/Inventory.cs
@@ -10,7 +10,7 @@ namespace Transportation
         [Required]
         public long ProductID { get; set; }
         public virtual Product Product { get; set; }
-        public long Quantity { get; set; }
+        public double Quantity { get; set; }
         public long LatestPrice { get; set; }
 
         public Inventory() {
@@ -37,7 +37,7 @@ namespace Transportation
         {
             ID = json.Value<long>("id");
             ProductID = json.Value<long>("productId");
-            Quantity = json.Value<long>("quantity");
+            Quantity = json.Value<double>("quantity");
             LatestPrice = json.Value<long>("latestPrice");
         }
     }
